pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helm-commander"
version = "0.1.0"
description = "Better visibility into Helm deployments on Kubernetes clusters"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "typer[all]>=0.12.0",
    "kubernetes>=28.0.0",
    "rich>=13.7.0",

    "pyyaml>=6.0",
    "deepdiff>=7.0",
    "packaging>=23.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
hcom = "helm_commander.cli.app:main"
helm-commander = "helm_commander.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/helm_commander"]
